task-3-kafka/producer/requirements.txt
fastapi
uvicorn[standard]
kafka-python-ng
pydantic